move decompressPng code and shim it on nodejs
This commit makes sure that
- in a browser a canvas element will be used for decoding pngs (no
  external dependencies needed)
- in node it will use the pngparse module which will use the native zlib
  module to inflate the data
