+++ package.json
@@ -25,7 +25,6 @@
   },
   "dependencies": {
     "aliasify-patch": "^1.4.1",
-    "canvas": "~1.2.8",
     "eventemitter2": "~0.4.13",
     "object-assign": "^2.0.0",
     "pngparse": "^2.0.1",
@@ -42,7 +41,8 @@
       "canvas": "./src/util/shim/canvas.js",
       "eventemitter2": "./src/util/shim/EventEmitter2.js",
       "ws": "./src/util/shim/WebSocket.js",
-      "xmldom": "./src/util/shim/xmldom.js"
+      "xmldom": "./src/util/shim/xmldom.js",
+      "../util/decompressPng": "./src/util/shim/decompressPng"
     }
   },
   "directories": {
